+++ threedi_modelchecker/config.py
@@ -28,10 +28,10 @@ from .checks.other import ConnectionNodesLength
 from .checks.other import CrossSectionLocationCheck
 from .checks.other import OpenChannelsWithNestedNewton
 from .checks.other import Use0DFlowCheck
-from .checks.timeseries import TimeseriesFloatCheck
 from .checks.timeseries import TimeseriesIncreasingCheck
-from .checks.timeseries import TimeseriesIntCheck
 from .checks.timeseries import TimeseriesRowCheck
+from .checks.timeseries import TimeseriesTimestepCheck
+from .checks.timeseries import TimeseriesValueCheck
 from .threedi_model import models
 from .threedi_model.models import constants
 from geoalchemy2 import functions as geo_func
@@ -1321,7 +1321,7 @@ CHECKS += [
     ]
 ]
 CHECKS += [
-    TimeseriesIntCheck(col, error_code=1201)
+    TimeseriesTimestepCheck(col, error_code=1201)
     for col in [
         models.BoundaryCondition1D.timeseries,
         models.BoundaryConditions2D.timeseries,
@@ -1330,7 +1330,7 @@ CHECKS += [
     ]
 ]
 CHECKS += [
-    TimeseriesFloatCheck(col, error_code=1202)
+    TimeseriesValueCheck(col, error_code=1202)
     for col in [
         models.BoundaryCondition1D.timeseries,
         models.BoundaryConditions2D.timeseries,